pyproject.toml
[project]
name = "pdf-chatbot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.135.1",
    "inngest>=0.5.18",
    "langchain>=1.2.13",
    "langchain-community>=0.4.1",
    "langchain-ollama>=1.0.1",
    "llama-index-core>=0.14.18",
    "llama-index-readers-file>=0.6.0",
    "ollama>=0.6.1",
    "openai>=2.29.0",
    "python-dotenv>=1.2.2",
    "qdrant-client>=1.17.1",
    "streamlit>=1.55.0",
    "transformers>=5.3.0",
    "uvicorn>=0.42.0",
]
